tools/requirements.txt
pyserial>=3.5
websocket-client>=1.7.0
python-dotenv>=1.0.0